+++ R/plotBuilder_la.R
@@ -84,9 +84,10 @@ if(ci==FALSE){
                                      layout.pos.col = matchidx$col))
      }
     }else{
-      for(i in 1:numPlots){
-        print(Plot[[i]])
-      }
+      all_plots <- lapply(1:numPlots, function(x) Plot[[x]])
+      ml <- gridExtra::marrangeGrob(all_plots, nrow = 1, ncol = 1,
+                                    top = " ")
+      invisible(print(ml))
     }
   }
 }else{
@@ -160,9 +161,10 @@ if(ldb == 1) {
                                    layout.pos.col = matchidx$col))
         }
   }else{
-    for(i in 1:numPlots){
-      print(Plot[[i]])
-        }
+    all_plots <- lapply(1:numPlots, function(x) Plot[[x]])
+    ml <- gridExtra::marrangeGrob(all_plots, nrow = 1, ncol = 1,
+                                  top = " ")
+      invisible(print(ml))
       }
     }
   }